Rename files and replace calls by services
